Refactored package name and changed app ID
Refactored package name from michiganhackersapp to michiganhackers.
Also changed app ID to michiganhackers.
https://developer.android.com/studio/build/application-id makes it seem
as though refactoring the package name is not necessary, but I did it
just in case because the google calendar api needs a client ID based off
the package name.
EDIT: Refactoring the name did not remove the org prefix. I am not sure if google api will work with this or not yet.