Currently working on Change Email. Not done.
Deleted account activity
Added signout button to settings
Created FirebaseAuthActivity that can be extended whenever we have an activity that needs to make sure the current user is logged in.
Split up profileviewmodel so that we can just use the member live data or teams live data logic in other parts of the app.
Reset pwd activity now signs out.