Added alert dialog and snackbar themes
Wanted to do a lighter background color and black text color for snackbar in dark theme, but android:textColor attribute wasn't changing text color. So, I just settled on a dark orange background. Snackbar theme isn't changed for light theme, but I added it to light theme style file to make it easier to change in the future.